chore: Move all `TileManifest` methods into .cc
